+++ src/main/java/aeronicamc/mods/mxtune/gui/widget/MXButton.java
@@ -126,7 +126,7 @@ public class MXButton extends Button implements ILayout, IHooverText
 
     @Override
     public boolean isMouseOver(double pMouseX, double pMouseY) {
-        return this.visible && pMouseX >= (double)this.x && pMouseY >= (double)this.y && pMouseX < (double)(this.x + this.width) && pMouseY < (double)(this.y + this.height);
+        return this.visible && pMouseX >= this.x && pMouseY >= this.y && pMouseX < (this.x + this.width) && pMouseY < (this.y + this.height);
     }
 
     @Override
